fix and test importing
